refactor: 적절한 폴더링을 진행한다
